Add support for running on macOS and Windows
Prior to this change, this action was a Docker-based action (i.e., a
"container action"). As seen in #2, container actions only run on Linux.

In this commit, we switch to using a "composite action", which allows us
to run on all three of the operating systems supported by GitHub
Actions: Linux, macOS, and Windows.